Fix problem related to zwj characters in emojis
This fixes a problem with the zwj character and headings which contain emoji sequences.

An example of the bug can be seen in the [👷🏼‍♀️ Mainenance link](https://github.com/neontribe/www/tree/0b8a8c85bcb9c39018ddcfc24df28879c10c81a0#neontribe-website). 

Github seems to leave the `\u200D` chracter un-encoded in the links, so when you click the `[👷🏼‍♀️ Maintenance](#%E2%80%8D-maintenance)` link above it doesn't jump to the correct place in the page.

This fix just makes sure that it works on github, I haven't tested it on the other MD renderers, so I was more defensive with the change.