Upgrade `emoji-regex` and `tap`, enhance test coverage
`emoji-regex` is very out of date at this point, and the old version
doesn't properly process many recent emoji additions.

I added a few tests for those recent additions that now pass, but did not before.

`tap`, I figured, it would just be good to update - and it was a non-breaking change, which is nice.